refactor(sdk-core): tss refactor
Ticket: BG-49287
