+++ modules/sdk-core/src/bitgo/tss/eddsa/index.ts
@@ -1,0 +1,11 @@
+export {
+  createCombinedKey,
+  createUserSignShare,
+  createUserToBitGoGShare,
+  offerUserToBitgoRShare,
+  getBitgoToUserRShare,
+  sendUserToBitgoGShare,
+  getTxRequest,
+  sendSignatureShare,
+  encryptYShare,
+} from './eddsa';
