+++ modules/sdk-core/src/bitgo/tss/index.ts
@@ -1,0 +1,13 @@
+export { EncryptedYShare, DecryptableYShare, CombinedKey, SigningMaterial, ShareKeyPosition } from './types';
+
+export {
+  createCombinedKey,
+  createUserSignShare,
+  createUserToBitGoGShare,
+  offerUserToBitgoRShare,
+  getBitgoToUserRShare,
+  sendUserToBitgoGShare,
+  getTxRequest,
+  sendSignatureShare,
+  encryptYShare,
+} from './eddsa';
